So much stuff
added client mailroom and engine, almost(maybe have) finished client
engine, starting server mailroom, established most of the communication
logic, updated player and cleaned some code

+++ src/Actors/Mob.java
@@ -3,7 +3,7 @@ import Util.StdDraw;
 public class Mob extends Actor {
      int hp;
      double offset = 0.35;
-     public Mob(int x, int y, int r, int hp) {
+     public Mob(double x, double y, int r, int hp) {
           super(x, y, r);
           this.hp = hp;
      }
